+++ source/Nevermore.Tests/Util/DataModificationQueryBuilderFixture.InsertMultipleDocumentWithMultipleRelatedDocumentsMaps.approved.txt
@@ -1,16 +1,12 @@
 INSERT INTO [dbo].[TestDocumentTbl]  ([Id], [AColumn], [JSON])  VALUES 
 (@Id, @AColumn, @JSON)
-INSERT INTO [dbo].[RelatedDocument] ([Id], [Table], [RelatedDocumentId], [RelatedDocumentTable]) VALUES
-(@Id, 'TestDocumentTbl', @relateddocument_0, 'OtherTbl')
-,(@Id, 'TestDocumentTbl', @relateddocument_1, 'OtherTbl')
-INSERT INTO [dbo].[OtherRelatedTable] ([Id], [Table], [RelatedDocumentId], [RelatedDocumentTable]) VALUES
-(@Id, 'TestDocumentTbl', @otherrelatedtable_0, 'OtherTbl')
-,(@Id, 'TestDocumentTbl', @otherrelatedtable_1, 'OtherTbl')
+INSERT INTO [dbo].[RelatedDocument] ([Id], [Table], [RelatedDocumentId], [RelatedDocumentTable])
+SELECT [Id], [Table], [RelatedDocumentId], [RelatedDocumentTable] FROM @0__relatedDocumentTableValuedParameter
+INSERT INTO [dbo].[OtherRelatedTable] ([Id], [Table], [RelatedDocumentId], [RelatedDocumentTable])
+SELECT [Id], [Table], [RelatedDocumentId], [RelatedDocumentTable] FROM @1__relatedDocumentTableValuedParameter
 
 @Id=New-Id-1
 @JSON={}
 @AColumn=Doc1
-@relateddocument_0=Rel-1
-@relateddocument_1=Rel-2
-@otherrelatedtable_0=Rel-3-Other
-@otherrelatedtable_1=Rel-2
+@0__relatedDocumentTableValuedParameter=Nevermore.TableValuedParameter
+@1__relatedDocumentTableValuedParameter=Nevermore.TableValuedParameter
