+++ source/Nevermore.Tests/Util/DataModificationQueryBuilderFixture.UpdateWithManyRelatedDocuments.approved.txt
@@ -1,33 +1,20 @@
 UPDATE [dbo].[TestDocumentTbl]  SET [AColumn] = @AColumn, [JSON] = @JSON WHERE [Id] = @Id
 
-DECLARE @references as TABLE (Reference nvarchar(400), ReferenceTable nvarchar(400))
-
-DELETE FROM @references
-
-INSERT INTO @references VALUES (@relateddocument_0, 'OtherTbl'), (@relateddocument_1, 'OtherTbl')
-
 DELETE FROM [dbo].[RelatedDocument] WHERE [Id] = @Id
-    AND [RelatedDocumentId] not in (SELECT Reference FROM @references)
+    AND [RelatedDocumentId] not in (SELECT [RelatedDocumentId] FROM @0__relatedDocumentTableValuedParameter_)
 
 INSERT INTO [dbo].[RelatedDocument] ([Id], [Table], [RelatedDocumentId], [RelatedDocumentTable])
-SELECT @Id, 'TestDocumentTbl', Reference, ReferenceTable FROM @references t
-WHERE NOT EXISTS (SELECT null FROM [dbo].[RelatedDocument] r WHERE r.[Id] = @Id AND r.[RelatedDocumentId] = t.Reference )
-
-DELETE FROM @references
-
-INSERT INTO @references VALUES (@otherrelatedtable_0, 'OtherTbl'), (@otherrelatedtable_1, 'OtherTbl')
-
+SELECT [Id], [Table], [RelatedDocumentId], [RelatedDocumentTable] FROM @0__relatedDocumentTableValuedParameter_ t
+WHERE NOT EXISTS (SELECT null FROM [dbo].[RelatedDocument] r WHERE r.[Id] = @Id AND r.[RelatedDocumentId] = t.[RelatedDocumentId] )
 DELETE FROM [dbo].[OtherRelatedTable] WHERE [Id] = @Id
-    AND [RelatedDocumentId] not in (SELECT Reference FROM @references)
+    AND [RelatedDocumentId] not in (SELECT [RelatedDocumentId] FROM @1__relatedDocumentTableValuedParameter_)
 
 INSERT INTO [dbo].[OtherRelatedTable] ([Id], [Table], [RelatedDocumentId], [RelatedDocumentTable])
-SELECT @Id, 'TestDocumentTbl', Reference, ReferenceTable FROM @references t
-WHERE NOT EXISTS (SELECT null FROM [dbo].[OtherRelatedTable] r WHERE r.[Id] = @Id AND r.[RelatedDocumentId] = t.Reference )
+SELECT [Id], [Table], [RelatedDocumentId], [RelatedDocumentTable] FROM @1__relatedDocumentTableValuedParameter_ t
+WHERE NOT EXISTS (SELECT null FROM [dbo].[OtherRelatedTable] r WHERE r.[Id] = @Id AND r.[RelatedDocumentId] = t.[RelatedDocumentId] )
 
 @Id=Doc-1
 @JSON={}
 @AColumn=Doc1
-@relateddocument_0=Rel-1
-@relateddocument_1=Rel-2
-@otherrelatedtable_0=Rel-3-Other
-@otherrelatedtable_1=Rel-2
+@0__relatedDocumentTableValuedParameter_=Nevermore.TableValuedParameter
+@1__relatedDocumentTableValuedParameter_=Nevermore.TableValuedParameter
